+++ doc/CHANGELOG.rst
@@ -17,6 +17,9 @@ Latest Changes:
 
   - Fixed type mismatch in JavadocExtractor causing help() to fail on Java classes. #1149
 
+  - Fixed JArray constructor ignoring slice bounds when creating from sliced array. #845
+
+
 - **1.7.1 - 2026-05-06**
 
   - Binaries for OSX ARM64.